helm chart for PowerDNS deployment
This change has helm chart to deploy powerdns
service with sqlite3 backend. The base image
used is a sle15 based BCI image.
The following components of powerdns is installed:
authoritative server and recursor.
The authoritative server is running on port 54,
the API webserver running on port 8081 and
recursor running on port 53.

Currently we are using pdns 4.7.2 version.
The image supports enabling include-dir directive to
scan for additional configuration file for both auth
and recursor components which support customization
of configuration entries.
The customized configuration entries are loaded into
the container via the configmap
(pdns-additional-config-dir, pdns-recursor-config-dir)
based on the values provided in the values.yaml file.
The sqlite3 database is persisted on the local
storage on the host under /opt/powerdns.
There is init-db script which initializes the DB.
The configure.sh script creates the zone specified in the
values.yaml as part of the lifecycle postStart event if
the zone does not exist.

Signed-off-by: Meera Belur <mbelur@suse.com>