Enhancements to the media helm chart
Make the name of the persistentvolume a fixed value, independent of
the release name used to install the chart; this allows the chart to
be installed with arbitrary release names, though only allows one
instance per namespace.

Also add 'mediaVolume' settings to values.yaml allowing the location
of the media volume to be overridden.

Signed-off-by: Fergal Mc Carthy <fmccarthy@suse.com>