Initial metal3-deploy chart
Pulls in dependent charts, and sets up consistent overrides for their
relevant settings.

Currently the templates/NOTES.txt contains static content; this should
be expanded to report the deployed application/service endpoints as well
as identify next steps for the user to perform.

Add git ignorance for the Chart.lock file and charts/ subdirectory that
is dynamically created by the 'helm dependency update' command.

The README.md provides basic usage instructions.

The metal3-deploy/values.yaml provides a consistent example set of the
override settings for all of the subcharts it manages.

Signed-off-by: Fergal Mc Carthy <fmccarthy@suse.com>